refactor: Revert to original naming convention and GUID
