+++ NOTICE
@@ -1,4 +1,4 @@
-Splunk SOAR App: GroupIB_TI
+Splunk SOAR App: Group IB Threat Intelligence
 Copyright (c) Group-IB, 2026
 Third Party Software Attributions:
