+++ groupibthreatintelligenceandattribution_consts.py
